HACK: Fix Python initialization in vtkPVPythonCatalyst.
vtkPVPythonCatalyst module needs to init Python in static builds and hence needs
to depend on "all enabled modules". To do this cleanly, we decided to
add_subdirectory(...) for this module directly. That would work but there's 1
problem: VTKConfig.cmake builds up a VTK_MODULES_ENABLED variable that doesn't
end up including this vtkPVPythonCatalyst module. We have plans to fix VTK, but
until that happens, we process vtkPVPythonCatalyst as part of the module-top
loop (skipping the CMakeLists.txt) and then processing it again at a later time.

Change-Id: I1ad4465bb18423b3b06946db066ae6f1ee344132